requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
mysqlclient==2.2.7
python-decouple==3.8
Pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2